# perf(rust, python): improve performance for datetime parsing with %Z