Currently, `'%Z'` is either ignored within `strptime::parse`, or it just causes parsing to fallback to `NaiveDateTime::parse_from_str`

I'd suggest to just exclude it from `strptime::fmt_len` and let it go directly to the chrono parser

This comes with a little [perf improvement](https://www.kaggle.com/code/marcogorelli/polars-timing?scriptVersionId=121211658) for parsing local time zone names with `'%Z'`
```python
dates = pl.date_range(datetime(1, 1, 1), datetime(9999, 1, 1)).dt.strftime('%Y-%m-%d %H:%MUTC')

dates.str.strptime(pl.Datetime, '%Y-%m-%d %H:%M%Z')
```

Here:
- 3.02 s ± 24 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.02 s ± 42.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.01 s ± 38 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

upstream/master:
- 3.4 s ± 324 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.24 s ± 39.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.33 s ± 57.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)